# api/openapi.yaml
openapi: 3.1.0
info:
  title: AgentOS API
  version: 0.1.0
  summary: Multi-tenant AI-agent infrastructure platform API
  description: |
    AgentOS is a multi-tenant AI-agent infrastructure platform. This contract
    covers the versioned REST API (/api/v1), the worker pull endpoint, the
    Server-Sent Events (SSE) execution-event stream, and platform health /
    metrics endpoints.

    ## Authentication
    Two schemes are accepted on every protected endpoint (see
    `securitySchemes`):
    1. `bearerToken` — `Authorization: Bearer <token>` where `<token>` is the
       HMAC-SHA256 JWT-style token returned by `POST /auth/login`
       (registration creates the first OWNER user; log in to get a token).
    2. `apiKey` — `X-API-Key: ak_...` header. API keys are created server-side
       (dev key created at startup) and authenticate as the key's OWNER role.
    Browser `EventSource` clients that cannot set headers may pass
    `?api_key=<key>` as a query parameter (legacy compatibility).

    ## RBAC
    Permissions: `agents.read`, `agents.write`, `runs.read`, `runs.execute`,
    `users.manage`, `organization.manage`. Roles: `OWNER` (all), `ADMIN`
    (all but organization.manage), `MEMBER` (agents.read, runs.read,
    runs.execute), `VIEWER` (agents.read, runs.read). Each operation declares
    its required permission via `x-required-permission`.

    ## Multi-tenancy
    Requests are scoped to an organization derived from the auth token. For
...

    ## Error model
...

...
servers:
...
security:
- bearerToken: []
- apiKey: []
paths:
...
  /eval-datasets/{id}/run:
    post:
      operationId: runEvalDataset
      summary: Execute every case of a dataset against an agent (synchronous, bounded)
      description: |
        Runs each case through the agent runtime with a per-case context
        deadline (30s) and at most 50 cases. Returns the completed run id;
        fetch `GET /eval-runs/{id}` for results and summary. Cost is recorded
        as 0 until the runtime exposes cost (token usage only today).
      security:
      - bearerToken: []
      - apiKey: []
      tags:
      - Evaluations
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - agent_id
              properties:
                agent_id:
                  type: string
      responses:
        '200':
          description: Run finished; status is `completed` (per-case failures land in the results).
          content:
            application/json:
              schema:
                type: object
                required:
                - eval_run_id
                - status
                properties:
                  eval_run_id:
                    type: string
                  status:
                    type: string
                    enum:
                    - completed
                    - failed
        '400':
          description: Missing agent_id or validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvalErrorEnvelope'
        '404':
          description: Dataset or agent not found in the caller's organization.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvalErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          description: Evals runner not wired into the API server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvalErrorEnvelope'
...
  /runs:
    get:
      operationId: listRuns
      tags:
      - runs
      summary: List runs
      description: |
        Returns an envelope `{"runs":[...]}`. **Today this lists all runs
        across organizations** (handler TODO notes org filtering is future
        work). Unpaginated today.
      x-required-permission: runs.read
      x-unpaginated: true
      x-note: no organization filter applied yet
      responses:
        '200':
          description: Runs envelope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRunsResponse'
              example:
                runs:
                - ID: 3f6c2b62-2b57-4f8e-9e4e-2f4d3b6c9a10
                  OrganizationID: org-1
                  AgentID: agent-1
                  Input: What is 2+2?
                  Output: ''
                  Status: QUEUED
                  CreatedAt: '2025-01-01T12:00:00.123456789Z'
                  UpdatedAt: '2025-01-01T12:00:00.123456789Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createRun
      tags:
      - runs
      summary: Create and enqueue a run
      description: |
        Creates a run (status `QUEUED`) and enqueues an `agent.run` task.
        Responds 201 with the run ID and `status: "queued"`. Actual execution
        happens in the worker; watch `GET /runs/{id}/events` (SSE) for
        `status.changed` events.
      x-required-permission: runs.execute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRunRequest'
            example:
              organization_id: org-1
              agent_id: agent-1
              input: What is 2+2?
      responses:
        '201':
          description: Run created and task queued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRunResponse'
              example:
                run_id: 3f6c2b62-2b57-4f8e-9e4e-2f4d3b6c9a10
                status: queued
        '400':
          description: Invalid body or missing agent_id.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/LegacyTextError'
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /runs/{id}:
    get:
      operationId: getRun
      tags:
      - runs
      summary: Get a run by ID
      x-required-permission: runs.read
      parameters:
      - $ref: '#/components/parameters/RunID'
      responses:
        '200':
          description: The run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
              example:
                ID: 3f6c2b62-2b57-4f8e-9e4e-2f4d3b6c9a10
                OrganizationID: org-1
                AgentID: agent-1
                Input: What is 2+2?
                Output: The result of 2+2 is 4
                Status: COMPLETED
                CreatedAt: '2025-01-01T12:00:00.123456789Z'
                UpdatedAt: '2025-01-01T12:00:05.987654321Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /runs/{id}/steps:
    get:
      operationId: listRunSteps
      tags:
      - runs
      summary: List the execution steps (trace) of a run
      description: |
        Returns the recorded execution timeline for a run: one entry per model
        call and tool invocation, captured by the worker's runtime step
        recorder and persisted as `run_steps`. Each step carries its type,
        status, input/output metadata, error, token usage and timestamps.
        Steps are tenant-scoped: the org is derived from the run row.
      x-required-permission: runs.read
      parameters:
      - $ref: '#/components/parameters/RunID'
      responses:
        '200':
          description: Step list (possibly empty) for the run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunStepsResponse'
              example:
                run_id: run-1
                steps:
                - ID: 9e1f...
                  RunID: run-1
                  StepType: model
                  Status: succeeded
                  InputMeta:
                    input: What is 2+2?
                    name: openai-compatible
                    index: 1
                  OutputMeta:
                    output: '{"tool":"calculator","arguments":{"expression":"2+2"}}'
                  TokenUsage:
                    total_tokens: 42
                  StartedAt: '2025-01-01T12:00:00.5Z'
                  CompletedAt: '2025-01-01T12:00:01.1Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
...
components:
  parameters:
...
    RunID:
      name: id
      in: path
      required: true
      description: Run identifier (UUID v4 string).
      schema:
        type: string
      example: 3f6c2b62-2b57-4f8e-9e4e-2f4d3b6c9a10
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: HS256-HMAC
      description: |
        HMAC-SHA256 JWT-style token from `POST /auth/login` (24h expiry).
        Claims: user_id, organization_id, email, role, exp.
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Pre-provisioned API key (`ak_...`). Authenticates as the key's
        organization with the OWNER role. Browser EventSource clients may
        alternatively pass `?api_key=<key>` (legacy query-param auth).
  responses:
    Unauthorized:
      description: Missing/invalid credentials (plain text body today; JSON model is the target).
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/LegacyTextError'
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but lacking the required permission or organization access.
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/LegacyTextError'
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found.
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/LegacyTextError'
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    LegacyTextError:
      type: string
      description: Legacy plain-text error body emitted via http.Error during migration.
      example: agent not found
    Error:
      type: object
      description: Structured error model (internal/httpx.ErrorModel).
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable stable error code (e.g. unauthorized, forbidden, not_found,
                bad_request, internal_error).
              examples:
              - not_found
            message:
              type: string
              description: Human-readable message.
              examples:
              - agent not found
            request_id:
              type: string
              description: Correlation ID echoed from the X-Request-ID header (or generated per request).
              examples:
              - 1b8f0c47e51d2a3c4d5e6f708192a3b4
          required:
          - code
          - message
          - request_id
      required:
      - error
...
    Run:
      type: object
      description: |
        Agent execution record (runs.Run; no json tags — PascalCase keys).
        Status transitions: QUEUED → RUNNING → COMPLETED | FAILED.
      properties:
        ID:
          type: string
          description: UUID v4.
          examples:
          - 3f6c2b62-2b57-4f8e-9e4e-2f4d3b6c9a10
        OrganizationID:
          type: string
          examples:
          - org-1
        AgentID:
          type: string
          examples:
          - agent-1
        Input:
          type: string
          examples:
          - What is 2+2?
        Output:
          type: string
          description: Final output once the run completes (empty until then).
          examples:
          - The result of 2+2 is 4
        Status:
          type: string
          enum:
          - QUEUED
          - RUNNING
          - COMPLETED
          - FAILED
        CreatedAt:
          type: string
          format: date-time
        UpdatedAt:
          type: string
          format: date-time
      required:
      - ID
      - OrganizationID
      - AgentID
      - Status
      - CreatedAt
      - UpdatedAt
...
    CreateRunRequest:
      type: object
      properties:
        organization_id:
          type: string
          description: Optional; defaults to the authenticated principal's organization.
          examples:
          - org-1
        agent_id:
          type: string
          examples:
          - agent-1
        input:
          type: string
          description: Prompt/task input handed to the agent.
          examples:
          - What is 2+2?
      required:
      - agent_id
      - input
    CreateRunResponse:
      type: object
      properties:
        run_id:
          type: string
          description: 'Created run ID (fallback: queue task ID if the runs service is unavailable).'
          examples:
          - 3f6c2b62-2b57-4f8e-9e4e-2f4d3b6c9a10
        status:
          type: string
          const: queued
      required:
      - run_id
      - status
    ListRunsResponse:
      type: object
      properties:
        runs:
          type: array
          items:
            $ref: '#/components/schemas/Run'
      required:
      - runs
...
    RunStep:
      type: object
      description: |
        One recorded execution step of a run's timeline (runs.Step; no json
        tags — PascalCase keys). StepType is `model` or `tool`; Status is
        `succeeded`, `failed` or `PENDING`.
      properties:
        ID:
          type: string
        RunID:
          type: string
        StepType:
          type: string
          enum:
          - model
          - tool
          - PENDING
        Status:
          type: string
          enum:
          - succeeded
          - failed
          - PENDING
        InputMeta:
          type: object
          additionalProperties: true
          description: e.g. {input, name, index}
        OutputMeta:
          type: object
          additionalProperties: true
          description: e.g. {output}
        Error:
          type: string
        TokenUsage:
          type: object
          additionalProperties: true
          description: e.g. {prompt_tokens, completion_tokens, total_tokens}
        Cost:
          type: number
          format: double
        StartedAt:
          type: string
          format: date-time
        CompletedAt:
          type: string
          format: date-time
        CreatedAt:
          type: string
          format: date-time
    RunStepsResponse:
      type: object
      properties:
        run_id:
          type: string
        steps:
          type: array
          items:
            $ref: '#/components/schemas/RunStep'
    EvalErrorEnvelope:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
...